Remove unused file & function & update user/show
