--- resources/js/Shared/DisplayPageComponent/Element/Expandable.jsx
@@ -1,10 +1,0 @@
-import { useState } from "react";
-
-export default ({children}) => {
-    const [expand, setExpand] = useState(false);
-    
-    setExpand('expand');
-    return <div className={"expanable "+expand}>
-        {children}
-    </div>;
-}
