+++ resources/js/Shared/DisplayPageComponent/Form/CreateEditForm.jsx
@@ -1,0 +1,57 @@
+import React from 'react';
+import { useForm } from '@inertiajs/inertia-react';
+import LoadingButton from '@/Shared/LoadingButton';
+import InputField from "../Field/InputField";
+
+const CreateEditForm = ({Form, children}) => {
+    const method = Form.edit_form ? {_method: 'PUT'} : null;
+    const UseFormObject = {...Form.getVariableForUseForm(), ...method};
+    const { data, setData, errors, post, processing, transform } = useForm(UseFormObject);
+
+    window.data = data;
+    window.setData = setData;
+    window.tomatoDebugger = {
+        form : Form,
+        setData : setData
+    }
+    
+    transform((data) => {
+        Form.fields.filter(f => f.model).forEach(y => data[y.entityname] =  data[y.entityname].split(" : ")[0]);
+        return data;
+    })
+
+    function handleSubmit(e) {
+        e.preventDefault();
+        post(Form.submit_url);
+    }
+
+    return <div className="w-full">
+            <form onSubmit={handleSubmit}>
+                <div className="flex flex-wrap p-6 -mb-8 -mr-6">
+                    {Form.fields.map((Field, keyId) => {
+                        return <InputField 
+                            Field={Field} 
+                            key={keyId}
+                            errors={errors}
+                            data={data}
+                            setData={setData}    
+                        ></InputField>
+                    })}
+                </div>
+
+                {children}
+
+                <div className="flex items-center px-8 py-4 bg-gray-100 border-t border-gray-200">
+                    <LoadingButton
+                        loading={processing}
+                        type="submit"
+                        className="ml-auto btn-indigo"
+                    >
+                        Save
+                    </LoadingButton>
+                </div>
+            </form>
+        </div>
+}
+
+export default CreateEditForm;
