+++ resources/js/Shared/DisplayPageComponent/Form/DisplayForm.jsx
@@ -1,16 +1,12 @@
-import { getAlias } from "@/util";
-import { FieldUtil } from "../util_form";
-import { Field } from "./util_field";
-
-
-export default ({fields, form}) => {
+import { FieldUtil } from "../Field/util_field";
+export default ({fields}) => {
     if(fields){
         return fields.map((field,keyId) => {
-            return FieldUtil.check_and_getCommonField(field,keyId) || 
+            return FieldUtil.getElementIfExist(field,keyId) || 
                         (<div key={keyId} className="form-component pr-6 pb-8 w-full lg:w-1/2">
                             <label className="form-label">{field.label}:</label> 
                             <div className="form-input">
-                                {Field.displayElement(field)}
+                                {FieldUtil.getProcessedContent(field)}
                             </div>
                         </div>);
         })
