+++ resources/js/Shared/DisplayPageComponent/Form/util_form.js
@@ -1,5 +1,3 @@
-import React from "react";
-
 export class FormSchema {
     constructor(form_schema){
         this.schema = {...form_schema};
@@ -7,6 +5,7 @@ export class FormSchema {
         this.modal = this.schema.modal;
         this.id = this.schema.id;
         this.fields = this.schema.fields;
+
         this.fields_has_entity = this.fields.filter(field => {
             return field.entityname
         });
@@ -24,18 +23,5 @@ export class FormSchema {
 
     getVariableForUseForm(){
         return this.fields.reduce((obj, item) => Object.assign(obj, { [item.entityname] : (item.value || '') }), {});
-    }
-}
-
-export class FieldUtil{
-    static check_and_getCommonField(field, keyId){
-        if(field.element){
-             if(field.element == 'row'){
-                return React.createElement('div', {className:'row w-full', key:keyId}, '');
-            }else if(field.element == 'line'){
-                return React.createElement('div', {className:'line w-full', key:keyId}, '');
-            }
-        }
-        return null;
     }
 }
