+++ resources/js/Shared/LayoutComponent/HeaderAction.jsx
@@ -1,9 +1,9 @@
-import { CurrentUser } from "@/util";
+import { CurrentUser } from "../Util";
 
 export default ({action}) => {
     const user = new CurrentUser();
     const model = window.location.pathname.split('/')[1];
-
+    
     return (<span className="header-action">
         {
             user.hasPermission('create-'+model) && 
@@ -11,20 +11,3 @@ export default ({action}) => {
         }
     </span>);
 }
-
-/*
-        {
-            user.hasPermission('create-user') && actions.includes('create-user') && 
-            
-        }
-
-        {
-            user.hasPermission('create-branch') && actions.includes('create-branch') && 
-            <a href="/branch/create" className="button btn-sm btn-light">Create New Branch</a>
-        }
-
-        {
-            user.hasPermission('create-employee') && actions.includes('create-employee') && 
-            <a href="/employee/create" className="button btn-sm btn-light">Create Employee</a>
-        }
-*/
