--- resources/js/Shared/unused/RouteLink.jsx
@@ -1,13 +1,0 @@
-import { InertiaLink } from "@inertiajs/inertia-react"
-
-export default ({ob, routename}) => {
-    return <>
-        <InertiaLink
-            tabIndex="-1"
-            // href={route(routename, ob.id)}
-            className="link"
-        >
-            {ob.name}
-        </InertiaLink>
-    </>
-}
